Control PIR sensor with OLED module
Attempt at making a somewhat interactable smart alarm. The OLED 1.3 module has two buttons, and I built a primitive interface with them to turn the alarm on and off by disabling the PIR sensor. A nice HAT serves as the "module".

TBD: add appropriate alarm functions with lights and alarm speakers, replace the standard Pico with a Pico W to send wireless signals about the alarm and to toggle the alarm system on/off.